Make roads installable
