+++ pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "road_flow_modelling"
+version = "0.1.0"
